L7 L8 - Api and Front End/ai-hedgefund-trading-system/requirements.txt
pandas
numpy
ccxt
ta
flask
flask-cors
torch
scikit-learn
stable-baselines3
sqlalchemy
websocket-client